refactor(connect): move solana method definitions to corresponding folder
