+++ packages/connect-common/src/types/api/solana/solanaComposeTransaction.ts
@@ -1,0 +1,6 @@
+import type { SolanaComposeTransaction, SolanaComposedTransaction } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function solanaComposeTransaction(
+    params: Params<SolanaComposeTransaction>,
+): Response<SolanaComposedTransaction>;
