+++ packages/connect-common/src/types/api/solana/solanaGetAddress.ts
@@ -1,4 +1,4 @@
-import type { Address, BundledParams, GetAddress, Params, Response } from '../params';
+import type { Address, BundledParams, GetAddress, Params, Response } from '../../params';
 
 export declare function solanaGetAddress(params: Params<GetAddress>): Response<Address>;
 export declare function solanaGetAddress(params: BundledParams<GetAddress>): Response<Address[]>;
