+++ packages/connect-common/src/types/api/solana/solanaGetPublicKey.ts
@@ -1,5 +1,5 @@
-import type { BundledParams, GetPublicKey, Params, Response } from '../params';
-import type { SolanaPublicKey } from './solana';
+import type { SolanaPublicKey } from './common';
+import type { BundledParams, GetPublicKey, Params, Response } from '../../params';
 
 export declare function solanaGetPublicKey(params: Params<GetPublicKey>): Response<SolanaPublicKey>;
 export declare function solanaGetPublicKey(
