+++ packages/connect-common/src/types/api/solana/solanaSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { SolanaSignTransaction, SolanaSignedTransaction } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function solanaSignTransaction(
+    params: Params<SolanaSignTransaction>,
+): Response<SolanaSignedTransaction>;
