--- packages/connect-common/src/types/api/solanaComposeTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { SolanaComposeTransaction, SolanaComposedTransaction } from './solana';
-
-export declare function solanaComposeTransaction(
-    params: Params<SolanaComposeTransaction>,
-): Response<SolanaComposedTransaction>;
