--- packages/connect-common/src/types/api/solanaSignTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { SolanaSignTransaction, SolanaSignedTransaction } from './solana';
-
-export declare function solanaSignTransaction(
-    params: Params<SolanaSignTransaction>,
-): Response<SolanaSignedTransaction>;
