+++ packages/connect-common/src/types/index.ts
@@ -14,7 +14,7 @@ export * from './api/bitcoin/common';
 export * from './api/ripple/common';
 export * from './api/ethereum/common';
 export * from './api/monero/common';
-export * from './api/solana';
+export * from './api/solana/common';
 export * from './api/stellar';
 export * from './api/tezos';
 export * from './api/tron';
